+++ project_cylon/CommonSteps.py
@@ -1,271 +1,431 @@
 # -*- coding: utf-8 -*-
+
 from behave import *
-from project_cylon.WorldContext import *
 
-##
-## Given step definitions
-##
+from Logger import *
+from CustomOperators import *
 
-@step ("User has [{PageName}/{PathName}] page opened") ##->given
+from WorldContext import *
+
+
+"""
+For create combination steps
+"""
+
+@step ("{Execution}, accept fail")
+def step_impl(context, Execution):
+    try:
+        context.execute_steps(u"""When %s""" % Condition)
+    except: pass
+
+
+@step ("{Condition}, {Execution}")
+def step_impl(context, Condition, Execution):
+    try:
+        runResult = False
+        context.execute_steps(u"""When %s""" % Condition)
+        runResult = True
+    except: pass
+
+    if runResult == True:
+        context.execute_steps(u"""Then %s""" % Execution)
+
+
+# @step ("[remote] {Statement}")
+# def step_impl(context, Statement):
+#     World.driver = World.secondary_driver
+#     context.execute_steps(u"""When %s""" % Statement)
+#     World.driver = World.primary_driver
+
+
+"""
+Given step definitions
+"""
+
+@step ("user has [{PageName}/{PathName}] page opened") ##->given
 def step_impl(context, PageName, PathName):
     Page = World.FindPage(PageName)
     Page.Go(PathName)
 
 
-#@step ("ผู้ใช้เปิดหน้าเว็บ [{PageName}]")
-@step ("User has [{PageName}] page opened") ##->given
-@step ("User has [{PageName}] page open")
-@step ("User has [{PageName}] open")
+@step ("user has [{PageName}] page opened") ##->given
+@step ("user has [{PageName}] page open")
+@step ("user has [{PageName}] open")
 def step_impl(context, PageName):
     Page = World.FindPage(PageName)
     Page.Go()
 
 
-@step ("User browse to url '{URL}'")
+@step ("user browse to url '{URL}'") ##->given
 def step_impl(context, URL):
     Page = World.CreateDummyPage(URL)
     Page.Go()
 
 
-##
-## When step definitions
-##
-#@step ("ผู้ใช้กรอกข้อมูล '{Value}' ลงใน [{ElementName}]")
-@step ("User enters '{Value}' to the [{ElementName}]") ##->when
-@step ("User enters '{Value}' to [{ElementName}]")
+"""
+When step definitions
+"""
+
+@step ("user enters '{Value}' to the [{ElementName}]") ##->when
+@step ("user enters '{Value}' to [{ElementName}]")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
     Element.SendKeys(Value)
 
 
-@step ("User enters date '{Value}' to the [{ElementName}]") ##->when
-@step ("User enters date '{Value}' to [{ElementName}]")
+@step ("user enters following lines to the [{ElementName}]")
+def step_impl(context, ElementName):
+    Element = World.FindElement(ElementName)
+    Element.SendKeys(context.text)
+
+
+@step ("user enters date '{Value}' to the [{ElementName}]") ##->when
+@step ("user enters date '{Value}' to [{ElementName}]")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
     Element.SendKeysByScript(Value)
 
 
-@step ("User clears value on the [{ElementName}]") ##->when
-@step ("User clears input [{ElementName}]")
+@step ("user clears value on the [{ElementName}]") ##->when
+@step ("user clears input [{ElementName}]")
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
     Element.SendKeysByScript("")
 
 
-#@step ("ผู้ใช้คลิกเลือก [{ElementName}]")
-@step ("User selects the [{ElementName}]") ##->when
-@step ("User clicks the [{ElementName}]") ##->when
-@step ("User clicks [{ElementName}] link")
-@step ("User clicks [{ElementName}] button")
+@step ("user selects the [{ElementName}]") ##->when
+@step ("user clicks the [{ElementName}]") ##->when
+@step ("user clicks [{ElementName}] link")
+@step ("user clicks [{ElementName}] button")
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
     Element.Click()
 
 
-#@step ("ผู้ใช้เลือก '{Value}' ใน [{ElementName}]")
-@step ("User selects '{Value}' on the [{ElementName}]") ##->when
-@step ("User selects '{Value}' in [{ElementName}]")
+@step ("user selects '{Value}' on the [{ElementName}]") ##->when
+@step ("user selects '{Value}' in [{ElementName}]")
 def step_impl(context, Value, ElementName):
     Element = World.FindElement(ElementName)
     Element.Select(Value)
 
 
-@step ("User checks the [{ElementName}]") ##->when
+@step ("user checks the [{ElementName}]") ##->when
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
     Element.Check()
 
 
-@step ("User unchecks the [{ElementName}]") ##->when
+@step ("user unchecks the [{ElementName}]") ##->when
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
     Element.Uncheck()
 
 
-@step ("User moves mouse over the [{ElementName}]") ##->when
-@step ("User moves mouse over [{ElementName}]")
+@step ("user moves mouse over the [{ElementName}]") ##->when
+@step ("user moves mouse over [{ElementName}]")
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
     Element.MouseOver()
 
 
-@step ("User uploads file '{FilePath}' to the [{ElementName}]") ##->when
-@step ("User uploads file '{FilePath}' to [{ElementName}]")
+@step ("user uploads file '{FilePath}' to the [{ElementName}]") ##->when
+@step ("user uploads file '{FilePath}' to [{ElementName}]")
 def step_impl(context, FilePath, ElementName):
     Element = World.FindElement(ElementName)
     Element.SendKeys(FilePath)
 
 
 #@step ("ผู้ใช้กรอกข้อมูล '{Value}' ลงในหน้าต่างแจ้งเตือน")
-@step ("User enters '{Value}' to the popup")
+@step ("user enters '{Value}' to the popup")
 def step_impl(context, Value):
     World.SendKeysToPopup(Value)
 
 
-@step ("User accept the popup")
-@step ('User clicks "OK" on the popup')
+@step ("user accept the popup")
+@step ('user clicks "OK" on the popup')
 def step_impl(context):
     World.AcceptPopup()
 
 
-@step ("User cancel the popup")
-@step ('User clicks "Cancel" on the popup')
+@step ("user cancel the popup")
+@step ('user clicks "Cancel" on the popup')
 def step_impl(context):
     World.DismissPopup()
 
-##
-## Then step definitions
-##
-@step ("The browser shows [{PageName}/{PathName}] page") ##->then
+"""
+Then step definitions
+"""
+
+@step ("the browser shows [{PageName}/{PathName}] page") ##->then
 def step_impl(context, PageName, PathName):
     Page = World.FindPage(PageName)
     Page.VerifyPage(PathName)
 
 
-@step ("The browser shows [{PageName}] page") ##->then
-@step ("The system displays [{PageName}]")
+@step ("the browser shows [{PageName}] page") ##->then
+@step ("the system displays [{PageName}]")
 def step_impl(context, PageName):
     Page = World.FindPage(PageName)
     Page.VerifyPage()
 
 
-@step ("The page url is '{URL}'") ##->then
-@step ("The system URL is '{URL}'")
-@step ("The system displays '{URL}'")
+@step ("the page url is '{URL}'") ##->then
+@step ("the system URL is '{URL}'")
+@step ("the system displays '{URL}'")
 def step_impl(context, URL):
     World.VerifyURLIs(URL)
 
 
-@step ("The page url contains '{URL}'") ##->then
-@step ("The system URL contains '{URL}'")
+@step ("the page url contains '{URL}'") ##->then
+@step ("the system URL contains '{URL}'")
 def step_impl(context, URL):
     World.VerifyURLContains(URL)
 
 
-@step ("The [{ElementName}] value is '{Value}'") ##->then
-@step ("The [{ElementName}] shows '{Value}'")
+@step ("the [{ElementName}] value is '{Value}'") ##->then
+@step ("the [{ElementName}] shows '{Value}'")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueIs(Value)
+
+    if Element.Value == Value:
+        return True
+    else:
+        Log.Failed("Verify value is?", Element.Value, Value)
 
 
-@step ("The [{ElementName}] value is not '{Value}'") ##->then
+@step ("the [{ElementName}] value is") ##->then
+def step_impl(context, ElementName):
+    Element = World.FindElement(ElementName)
+
+    if Element.Value == context.text:
+        return True
+    else:
+        Log.Failed("Verify value is? (multi-line)", Element.Value, context.text)
+
+
+@step ("the [{ElementName}] value is not '{Value}'") ##->then
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueIsNot(Value)
+
+    if Element.Value != Value:
+        return True
+    else:
+        Log.Failed("Verify value is not?", Element.Value, Value)
 
 
-@step ("The [{ElementName}] value contains '{Value}'") ##->then
-@step ("The [{ElementName}] contains '{Value}'")
+@step ("the [{ElementName}] value contains '{Value}'") ##->then
+@step ("the [{ElementName}] contains '{Value}'")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueContains(Value)
+
+    if Value in Element.Value:
+        return True
+    else:
+        Log.Failed("Verify value contains?", Element.Value, Value)
 
 
-@step ("The [{ElementName}] value is more than '{Value}'") ##->then
-@step ("The [{ElementName}] more than '{Value}'")
+@step ("the [{ElementName}] value contains") ##->then
+def step_impl(context, ElementName):
+    Element = World.FindElement(ElementName)
+
+    if context.text in Element.Value:
+        return True
+    else:
+        Log.Failed("Verify value contains? (multi-line)", Element.Value, Value)
+
+
+@step ("the [{ElementName}] value is more than '{Value}'") ##->then
+@step ("the [{ElementName}] more than '{Value}'")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueMoreThan(Value)
+
+    if Element.Value |more_than| Value:
+        return True
+    else:
+        Log.Failed(
+            "Verify value is more than?",
+            "value = %s" % Element.Value,
+            "value > %s" % Value
+        )
 
 
-@step ("The [{ElementName}] value is more than or equal '{Value}'") ##->then
-@step ("The [{ElementName}] more than or equal '{Value}'")
+@step ("the [{ElementName}] value is more than or equal '{Value}'") ##->then
+@step ("the [{ElementName}] more than or equal '{Value}'")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueMoreThanOrEqual(Value)
+
+    if Element.Value |more_than_or_equal| Value:
+        return True
+    else:
+        Log.Failed(
+            "Verify value is more than or equal?",
+            "value = %s" % Element.Value,
+            "value >= %s" % Value
+        )
 
 
-@step ("The [{ElementName}] value is less than '{Value}'") ##->then
-@step ("The [{ElementName}] less than '{Value}'")
+@step ("the [{ElementName}] value is less than '{Value}'") ##->then
+@step ("the [{ElementName}] less than '{Value}'")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueLessThan(Value)
+
+    if Element.Value |less_than| Value:
+        return True
+    else:
+        Log.Failed(
+            "Verify value is less than?",
+            "value = %s" % Element.Value,
+            "value < %s" % Value
+        )
 
 
-@step ("The [{ElementName}] value is less than or equal '{Value}'") ##->then
-@step ("The [{ElementName}] less than or equal '{Value}'")
+@step ("the [{ElementName}] value is less than or equal '{Value}'") ##->then
+@step ("the [{ElementName}] less than or equal '{Value}'")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueLessThanOrEqual(Value)
+
+    if Element.Value |less_than_or_equal| Value:
+        return True
+    else:
+        Log.Failed(
+            "Verify value is less than or equal?",
+            "value = %s" % Element.Value,
+            "value <= %s" % Value
+        )
 
 
-@step ("The [{ElementName}] value is between '{Value1}' and '{Value2}'") ##->then
+@step ("the [{ElementName}] value is between '{Value1}' and '{Value2}'") ##->then
 def step_impl(context, ElementName, Value1, Value2):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueBetween(Value1, Value2)
+
+    if (Element.Value |more_than_or_equal| Value1) and (Element.Value |less_than_or_equal| Value2):
+        return True
+    elif (Element.Value |less_than_or_equal| Value1) and (Element.Value |more_than_or_equal| Value2):
+        return True
+    else:
+        Log.Failed(
+            "Verify value is between?",
+            "value = %s" % Element.Value,
+            "value is between %s and %s" % (Value1, Value2)
+        )
 
 
-@step ("The [{ElementName}] value is empty") ##->then
-@step ("The [{ElementName}] is blank")
+@step ("the [{ElementName}] value is empty") ##->then
+@step ("the [{ElementName}] is blank")
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyValueIsBlank()
+
+    if Element.Value == "":
+        return True
+    else:
+        Log.Failed("Verify value is empty?", Element.Value, "<blank>")
 
 
-@step ("The [{ElementName}] value is not empty") ##->then
-@step ("The [{ElementName}] has a value")
+@step ("the [{ElementName}] value is not empty") ##->then
+@step ("the [{ElementName}] has a value")
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyHasValue()
+
+    if Element.Value != "":
+        return True
+    Log.Failed("Verify value is not empty?", Element.Value, "<any value>")
 
 
-@step ("The [{ElementName}] exists") ##->then
+@step ("the [{ElementName}] exists") ##->then
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyExists()
+
+    if Element.Exists:
+        return True
+    else:
+        Log.Failed("Verify element '%s' exists?" % Element.Name, "not exists", "exists")
 
 
-@step ("The [{ElementName}] does not exist") ##->then
+@step ("the [{ElementName}] does not exist") ##->then
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyNotExists()
+
+    if not Element.Exists:
+        return True
+    else:
+        Log.Failed("Verify element '%s' exists?" % Element.Name, "not exists", "exists")
 
 
-@step ("The [{ElementName}] is visible") ##->then
+@step ("the [{ElementName}] is visible") ##->then
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyVisible()
+
+    if Element.Visible:
+        return True
+    else:
+        Log.Failed("Verify element '%s' visible?" % Element.Name, "not visible", "visible")
 
 
-@step ("The [{ElementName}] is invisible") ##->then
+@step ("the [{ElementName}] is invisible") ##->then
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyNotVisible()
+
+    if not Element.Visible:
+        return True
+    else:
+        Log.Failed("Verify element '%s' not visible?" % Element.Name, "visible", "not visible")
 
 
-@step ("The [{ElementName}] is selected") ##->then
-@step ("The [{ElementName}] is checked")
+@step ("the [{ElementName}] is selected") ##->then
+@step ("the [{ElementName}] is checked") ##->then
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyIsChecked()
+
+    if Element.Selected:
+        return True
+    else:
+        Log.Failed("Verify element '%s' checked?" % Element.Name, "unchecked", "checked")
 
 
-@step ("The [{ElementName}] is not selected") ##->then
-@step ("The [{ElementName}] is unchecked")
+@step ("the [{ElementName}] is not selected") ##->then
+@step ("the [{ElementName}] is unchecked") ##->then
 def step_impl(context, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyIsUnchecked()
+
+    if not Element.Selected:
+        return True
+    else:
+        Log.Failed("Verify element '%s' unchecked?" % Element.name, "checked", "unchecked")
 
 
-#@step ("The [browser list] has item 'Opera'") :TODO
-
-
-@step ("The popup message shows '{Value}'") ##->then
-@step ("The popup shows '{Value}'")
+@step ("the popup message shows '{Value}'") ##->then
+@step ("the popup shows '{Value}'")
 def step_impl(context, Value):
-    World.VerifyPopupMessage(Value)
+    #World.VerifyPopupMessage(Value)
+    Message = World.GetPopupMessage()
+
+    if Message == Value:
+        return True
+    else:
+        Log.Failed("Verify popup message is?", Message, Value)
 
 
-@step ("The page has '{Amount}' items of [{ElementName}]") ##->then
-@step ("The system displays '{Amount}' items of [{ElementName}]")
+@step ("the page has '{Amount}' items of [{ElementName}]") ##->then
+@step ("the system displays '{Amount}' items of [{ElementName}]")
 def step_impl(context, Amount, ElementName):
     Element = World.FindElement(ElementName)
-    Element.VerifyItemsCount(Amount)
+
+    if Element.Count == int(Amount):
+        return True
+    else:
+        Log.Failed(
+            "Verify amount of '%s' is?" % Element.Name,
+            "amount = %s" % Element.Count,
+            "amount = %s" % Amount
+        )
 
 
-@step ("The [{ElementName}] tooltip text is '{Value}'") ##->then
-@step ("The [{ElementName}] tooltip is '{Value}'")
+@step ("the [{ElementName}] tooltip text is '{Value}'") ##->then
+@step ("the [{ElementName}] tooltip is '{Value}'")
 def step_impl(context, ElementName, Value):
     Element = World.FindElement(ElementName)
-    Element.VerifyTooltip(Value)
+
+    if Element.TooltipText == Value:
+        return True
+    else:
+        Log.Failed("Verify tooltip text is?", Element.TooltipText, Value)